remove: 불필요한 알림 컴포넌트 InspectionStatusAlarm 제거 (#27)
