+++ components/common/Alarm/styles.css.ts
